+++ packages/components-library/src/main.ts
@@ -9,6 +9,7 @@ import MultiFilter from './components/filters/MultiFilter.vue'
 import NumberFilter from './components/filters/NumberFilter.vue'
 import RangeFilter from './components/filters/RangeFilter.vue'
 import StringFilter from './components/filters/StringFilter.vue'
+import Toaster from './components/ui/Toaster.vue'
 import EntityTable from './components/table/EntityTable.vue'
 
 export {
@@ -23,5 +24,7 @@ export {
   NumberFilter,
   RangeFilter,
   StringFilter,
-  EntityTable
+  Toaster
+  EntityTable,
 }
+